refactor(f10): lift the nine dispatch config sections into core/
Audit F-10: runtime/config.h was included by 22 files in src/exec/, 85
translation units transitively, and changed 130 times in six months — the
highest build cost in the repo (#1222 measured it).

The nine sections src/exec reads (kv_cache, attention, moe, gdn, gemm,
generation, speculative, ffn, diagnostics) now live in src/core/dispatch_policy.h
under namespace imp::cfg, with DispatchPolicy aggregating them. Engine fills that
snapshot in init_weights() and hands exec a pointer to it instead of to the live
RuntimeConfig.

  exec/    -> runtime/config.h   22 -> 0
  compute/ -> runtime/config.h    2 -> 0
  src/ total                     33 -> 9, all nine now inside src/runtime/
  config.h                     1143 -> 403 lines
  TUs pulling config.h           85 -> 18

Three things the R-18 sketch did not cover, all found by building:

- struct and member are declared in one statement (`struct Diagnostics { ... }
  diagnostics;`), so a naive extraction takes the member with it and leaves
  RuntimeConfig fieldless.
- un-nesting collides: imp::KVCache is the paged cache in memory/kv_cache.h and
  imp-bench defines its own AttentionConfig. Hence namespace imp::cfg; member
  names are unchanged so every cfg.attention.x still reads the same.
- executor_workspace_buffers.cu const_casts the derived
  attention.fmha_prefill_threshold into the config AFTER handover — a write my
  earlier `runtime_config_.x.y =` grep missed. Checked: every reader of that
  field is inside src/exec/, so snapshot-writes-and-reads-itself preserves
  behaviour. A future mutation with readers outside exec would diverge silently,
  and dispatch_policy.h says so.

VALIDATED. A snapshot that drops a field throws no error, it picks a different
kernel and the test stays green — so the load-bearing check is not the test
count:

  Resolved-dispatch lines across the GPU suite: BIT-IDENTICAL, all twelve
  patterns at identical frequencies, before and after.

  test-e2e (excluding the 23 GB MtpForwardTest): 154/154, 0 OOM, 0 IMA.
  verify-fast: decode -1.49%, prefill +0.87%, pp4096 -0.41%, peak VRAM +0.01%,
  graphs 2.51x, degeneration smoke coherent. Spread 2.13% that run, so the
  decode delta is inside the noise, not evidence either way.
  ctest -L unit 4/4.

The full-suite run showed 26 failures and they are NOT this change: all of them
cascade from MtpForwardTest.DraftStepProducesValidToken, which loads a 23 GB
model on a 32 GB card and sometimes fails to release it (27 "out of memory"
lines). Two runs of the identical build differ there — pool drains in one
(reserved 23200->0) and freezes in the other (21856->21856) from a byte-identical
prior state, i.e. a capacity knife-edge, not determinism. Excluding that one test
the binary is clean.
