+++ src/exec/executor_ffn.cu
@@ -1,3 +1,4 @@
+#include "core/dispatch_policy.h"
 #include "exec/executor.h"
 #include "lora/lora_adapter.h"
 #include "exec/executor_kernels.h"
@@ -12,7 +13,6 @@
 #include "compute/ffn_sparsity_mask.h"
 #include "compute/ffn_sparsity_probe.h"
 #include "compute/hadamard.h"
-#include "runtime/config.h"
 #include "quant/quant_gemm.h"
 #include "quant/dequant_gpu.h"
 #include "quant/fp8_quant.h"
