+++ src/exec/executor_forward_moe_legacy.cu
@@ -1,12 +1,12 @@
 // Legacy MoE prefill fallback: D2H sync + serial/batch dequant + cuBLAS.
 // Extracted from executor_forward_moe.cu for maintainability.
 
+#include "core/dispatch_policy.h"
 #include "exec/executor.h"
 #include "exec/executor_forward_moe_internal.h"
 #include "exec/executor_kernels.h"
 #include "exec/gemm_context.h"
 #include "exec/executor_debug.h"
-#include "runtime/config.h"
 #include <atomic>
 #include "compute/embedding.h"
 #include "compute/ggml_mmvq.h"
