+++ src/exec/pre_dequant_phase3_fp8.cu
@@ -3,13 +3,13 @@
 // kernel file-size threshold. See pre_dequant_internal.h / quant_pipeline.h
 // for shared declarations.
 
+#include "core/dispatch_policy.h"
 #include "exec/executor.h"
 #include "exec/quant_pipeline.h"
 #include "exec/pre_dequant_internal.h"
 #include "quant/fp8_quant.h"
 #include "core/logging.h"
 #include "memory/vram_allocator.h"
-#include "runtime/config.h"
 
 #include <cuda_runtime.h>
 #include <cuda_fp16.h>
